+++ src/node/glossary.ts
@@ -1,7 +1,7 @@
 import type { PartialDeep } from 'type-fest'
-import type { HttpNetworkFrameEventMap } from '#core/future/frames/http-frame'
-import type { WebSocketNetworkFrameEventMap } from '#core/future/frames/websocket-frame'
-import type { AnyHandler } from '#core/future/handlers-controller'
+import type { HttpNetworkFrameEventMap } from '#core/experimental/frames/http-frame'
+import type { WebSocketNetworkFrameEventMap } from '#core/experimental/frames/websocket-frame'
+import type { AnyHandler } from '#core/experimental/handlers-controller'
 import type { LifeCycleEventEmitter, SharedOptions } from '#core/sharedOptions'
 
 export interface ListenOptions extends SharedOptions {}
